Add a two-part project_id by hand to the Stampede2 entries
This is two comma-separated values, the first one being the local allocation ID
on the resource, and the second is the OSG project name.

The CE job route will split these out and use them.

This will be automatically added by the frontend once SOFTWARE-4565 is ready;
in the meantime, set it by hand so we can test the job routes.