Add match expression logic to top-level of frontend.
As we're low in jobs, we're seeing a lot of requested glideins that are
not matchable, particularly because the remaining jobs are high-memory
(and they're causing pilots to be requested from low-memory slots).

Copied over the match expressions from CHTC.  Note they are python3
compatible and have some funkiness like:

```
int(glidein["attrs"]["GLIDEIN_MaxMemMBs"] or 0) == 0
```

This is to protect us from the fact the factory isn't consistent in
typing and sometimes sets attributes as strings and other times as
integers.

With this, a job can avoid a site by setting UNDESIRED_Sites; this
is significantly better than using the Requirements expression (as
glideinWMS understands this as well).