main 684: bring up to date with main-canary 684
Bring `ospool-pilot/main` up to date with `ospool-pilot/main-canary` 684,
which includes the `prepare-hook` hook as a replacement for `default_singularity_wrapper.sh`.

**prepare-hook**

The script `job/prepare-hook` is a "PREPARE_JOB" hook [1] [2].  It runs
before the job is started but after file transfer is complete; if it succeeds,
it modifies the job classad that the starter sees to set the variables that
cause the starter to invoke the job under Singularity (if it is available).
See the comments by `exit_hook()` for how errors work.

If the incoming job has an image path in the `SingularityImage` attribute,
we will download the image to a local (per-pilot) directory if necessary,
then rewrite `SingularityImage` to point to that file.  We also copy some
GlideinWMS files into the sandbox directory in the image.

**additional-htcondor-config**

This has been modified to set up the HTCondor knobs that tell HTCondor how
to invoke Singularity.  These knobs are set up before the pilot is started
and not per-job (as it was with the old job wrapper).  In particular:

- bind-mounts are detected up front; if a directory/file to be bind mounted
  does not exist on the host, that directory/file won't be attempted to be
  bind-mounted even if it's later created.

- The default singularity image is set at startup in
  `OSG_DEFAULT_SINGULARITY_IMAGE`; HTCondor is told to use this if
  `SingularityImage` is not found in the job ad.
  (The `pilot/default-image` script downloads this image up front if necessary.)

- Use of PID namespaces is set based on the `$SINGULARITY_USE_PID_NAMESPACES`
  environment variable at pilot startup; we use this instead of HTCondor's
  detection because some of the startd cron scripts need to know about it too.

- Set up the PREPARE_JOB hook.

See the logs and git blame for `ospool-pilot/itb` and `ospool-pilot/itb-canary` for a better story of the changes.

[1]: https://htcondor.readthedocs.io/en/latest/admin-manual/configuration-macros.html#configuration-file-entries-relating-to-job-hooks

[2]: https://htcondor.readthedocs.io/en/latest/admin-manual/hooks.html#job-hooks-that-fetch-work (search for `PREPARE_JOB`)