Do stashcp check in osgvo-node-advertise without modules
This changes the code for setting STASHCP_VERIFIED in
osgvo-node-advertise to just check for stashcp in PATH instead of trying
to load a very old version of stashcp via modules.  It also sets
STASHCP_VERSION based on the stashcp in PATH.

osgvo-node-advertise is still used by the gluex frontend, for example.