SOFTWARE-5896 Add development frontend group to ITB, move hello world
I didn't put all the guardrails up to prevent "devel" jobs from
landing on other groups' glideins, but for the lifetime of this
group, it will just need to be known that jobs need to be tagged
with MY.WantGlideinDevel = true and need to require
TARGET.IsGlideinDevel_Site.

This also pulls in and obsoletes PR #306, but places the hello
world script in the new group.